Removed inline CSS from footer and social HTML
I simply could not for the life of me figure out how to get both the social media icons and the copyright link to the site's license to align properly, with the buttons on the left and the link dead center. I also really don't like inline CSS, as it adds more work to maintain. So I just removed the inline CSS and kept the buttons on a seperate line. It's not the way I wanted, but it at least keeps the style on the site consistent by keeping everything centered.